requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
PyJWT==2.8.0
bcrypt==4.0.1
gunicorn==21.2.0
whitenoise==6.6.0